Specify install script

+++ persistent-hdfs/cleanup.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
+if false; then
+
 pushd /root > /dev/null
 
 rm hadoop-*.tar.gz > /dev/null 2>&1
 
 popd > /dev/null
+
+fi
